fix broken build

+++ dev/asset-bundle.cjs
@@ -5,7 +5,9 @@ const path = require('path');
 const rootDir = path.join(__dirname, '..');
 const buildDir = path.join(rootDir, 'build');
 
-fs.rmdirSync(buildDir, { recursive: true });
+if (fs.existsSync(buildDir)) {
+  fs.rmdirSync(buildDir, { recursive: true });
+}
 fs.mkdirSync(buildDir);
 for (const filename of ['favicon.ico', 'index.html']) {
   const src = path.join(rootDir, filename);
